[msx2] clean up mouse code
